refactor(tools): rename reply-react to react
PR #452 removed the message_reply tool; the file now only hosts
message_react. Drop the historical reply-react naming.

- src/tools/reply-react.ts → src/tools/react.ts
- src/tools/reply-react.test.ts → src/tools/react.test.ts
- ReplyReactToolContext → ReactToolContext
- createReplyReactTools → createReactTools
- logger tag tools:reply-react → tools:react

LazyTransportContext keeps its name — it's a generic transport holder,
not react-specific.

Pure rename, no behavior change.
